+++ .eslintrc.js
@@ -9,6 +9,8 @@ module.exports = {
   ],
   rules: {
     'no-param-reassign': ['error', { props: false }],
+    'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
+    'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off',
   },
   parserOptions: {
     parser: 'babel-eslint',
