+++ .gitignore
@@ -1,6 +1,5 @@
 .DS_Store
 node_modules
-/dist
 
 /tests/e2e/videos/
 /tests/e2e/screenshots/
