+++ src/api/server/index.js
@@ -3,11 +3,16 @@ import { http } from '@/plugins/http';
 export default {
   /**
    * Realiza um get para o endpoint configurado
-   * @param {string} endpoint
-   * @param {object} params
+   * @param {object} payload
    * @returns {Promise}
    */
-  checkUserStatus(endpoint, params) {
+  checkUserStatus(payload = {
+    endpoint: 'check-status',
+    identification: '',
+    identification_type: '',
+    store_id: '',
+  }) {
+    const { endpoint, ...params } = payload;
     return http.get(endpoint, { params }).then(response => response);
   },
 
@@ -17,17 +22,31 @@ export default {
    * @param {object} params
    * @returns {Promise}
    */
-  facebookLogin(endpoint, params) {
+  facebookLogin(payload = {
+    callback: '',
+    crossdm: '',
+    endpoint: 'facebook/url',
+    session_id: '',
+    store_id: '',
+  }) {
+    const { endpoint, ...params } = payload;
     return http.get(endpoint, { params }).then(response => response);
   },
 
   /**
    * Realiza um get para o endpoint configurado
    * @param {string} endpoint
-   * @param {object} params
+   * @param {object} payload
    * @returns {Promise}
    */
-  generateSecurityCode(endpoint, params) {
+  generateSecurityCode(payload = {
+    endpoint: 'generate-security-code',
+    identification: '',
+    identification_type: '',
+    session_id: '',
+    store_id: '',
+  }) {
+    const { endpoint, ...params } = payload;
     return http.post(endpoint, params).then(response => response);
   },
 
@@ -37,7 +56,13 @@ export default {
    * @param {object} params
    * @returns {Promise}
    */
-  hasAccount(endpoint, params) {
+  hasAccount(payload = {
+    identification: '',
+    endpoint: 'has-account',
+    session_id: '',
+    store_id: '',
+  }) {
+    const { endpoint, ...params } = payload;
     return http.get(endpoint, { params }).then(response => response);
   },
 
@@ -47,7 +72,14 @@ export default {
    * @param {object} params
    * @returns {Promise}
    */
-  passwordLogin(endpoint, params) {
+  passwordLogin(payload = {
+    identification: '',
+    endpoint: 'password',
+    session_id: '',
+    store_id: '',
+    identification_type: '',
+  }) {
+    const { endpoint, ...params } = payload;
     return http.post(endpoint, params).then(response => response);
   },
 };
