+++ src/plugins/index.js
@@ -1,7 +1,6 @@
-import httpPlugin from './http';
 import eventBus from './event-bus';
 
 export {
-  httpPlugin,
-  eventBus,
+  // eslint-disable-next-line
+  eventBus
 };
