+++ src/store/index.js
@@ -1,9 +1,9 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
 
-import state from './state';
-import mutations from './mutations';
 import actions from './actions';
+import mutations from './mutations';
+import state from './state';
 import * as getters from './getters';
 
 Vue.use(Vuex);
