+++ src/store/state.js
@@ -1,6 +1,3 @@
 export default {
-  errors: [],
   identification: '',
-  loading: false,
-  screen: 'Identification',
 };
